pyproject.toml
[project]
name = "challenge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.0",
    "pandas>=2.2.3",
    "pillow>=11.0.0",
]
